add protobuf decoder and encoder support

+++ rainbow-skt/src/main/java/com/milchstrabe/rainbow/skt/server/tcp/codc/annotion/NettyController.java
@@ -18,5 +18,5 @@ import java.lang.annotation.Target;
 @Component
 public @interface NettyController {
 
-    byte cmd();
+    int cmd();
 }
