+++ rainbow-skt/src/main/java/com/milchstrabe/rainbow/skt/server/tcp/codc/annotion/NettyMapping.java
@@ -15,5 +15,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface NettyMapping {
 
-    byte cmd();
+    int cmd();
 }
